LayoutTests/platform/mac/imported/w3c/web-platform-tests/content-security-policy/inheritance/inheritance-from-initiator.sub-expected.txt



PASS Setting src inherits from parent.
FAIL Changing contentWindow.location inherits from who changed it. assert_equals: expected "a" but got "p"
PASS Changing contentWindow.location indirectly inherits from who changed it directly.
FAIL window.open() inherits from caller. assert_equals: expected "a" but got "p"
PASS Click on anchor inherits from owner of the anchor.
FAIL Form submission through submit() inherits from owner of form. assert_equals: expected "b" but got "p"
FAIL Form submission through button click inherits from owner of form. assert_equals: expected "b" but got "p"
